major changes
- change congif struct
- extend command handler for word colour
- show red lights as long as there is no valid timestamp
- extend wordclock init
- new case for 1 am

+++ Firmware/wordclock/inc/eeprom.h
@@ -70,7 +70,7 @@
 #define PAGE_FULL             ((uint8_t)0x80)
 
 /* Variables' number */
-#define NB_OF_VAR             ((uint8_t)sizeof(config))
+#define NB_OF_VAR             ((uint16_t)sizeof(config))
 
 /* Exported types ------------------------------------------------------------*/
 /* Exported macro ------------------------------------------------------------*/
